pyproject.toml
[tool.poetry]
name = "reixi"
version = "6.0.0"
description = "utility-first general-purpose d*scord bot"
authors = ["Mark Joshwel <mark@joshwel.co>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/markjoshwel/reixi"
packages = [
    { include = "reixi" },
]
scripts = { reixi = "reixi.reixi:main" }

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
loguru = "^0.7.2"
uvloop = "^0.19.0"
tomlkit = "^0.12.3"
pydantic = "^2.5.3"
emoji = "^2.9.0"
thefuzz = "^0.20.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
mypy = "^1.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
